+++ paimon-spark/paimon-spark3-common/src/main/scala/org/apache/paimon/spark/catalyst/analysis/Spark3ResolutionRules.scala
@@ -50,7 +50,5 @@ case class Spark3ResolutionRules(session: SparkSession)
         .getOrElse {
           PaimonShowTablesExtendedCommand(catalog.asTableCatalog, ns, pattern, output)
         }
-
   }
-
 }
